feat(admin): assign org-group members from the Organizations panel
Surface org-group member assignment inside the platform Admin ->
Organizations panel, directly under the group-types card, so a platform
admin can grant a type and populate it without leaving the admin surface
to hand-navigate to /organizations/{id}.

Extract the reusable groups-with-members surface out of OrganizationGroups
into a shared OrganizationGroupsList, then consume it from both the
customer org page and a new admin OrganizationGroupMembersCard. The
assign/unassign endpoints already authorize platform admins, so this is a
UI-only change with no backend work.

Closes #1417
