+++ apps/client/app/components/admin/OrganizationProfileUpdated.tsx
@@ -23,6 +23,7 @@ import { getLastLoginDate } from '@client/app/utils/user';
 import Compressor from 'compressorjs';
 import { useUpdateOrganization } from '@client/app/hooks/data/organizations';
 import OrganizationGroupTypesCard from './OrganizationGroupTypesCard';
+import OrganizationGroupMembersCard from './OrganizationGroupMembersCard';
 
 export async function compressLogo(file: File, quality: number = 0.8): Promise<File | Blob> {
   return new Promise((resolve, reject) => {
@@ -213,6 +214,7 @@ const OrganizationProfileUpdated: React.FC<OrganizationProfileProps> = ({ org, o
               </Stack>
             </Card>
             <OrganizationGroupTypesCard org={org} />
+            <OrganizationGroupMembersCard org={org} />
             {children}
           </Grid>
         </Grid>
